+++ libtc6/inc/tc6.h
@@ -55,7 +55,7 @@ extern "C" {
 #define TC6_LIB_VER_MAJOR  (3U)
 #define TC6_LIB_VER_MINOR  (1U)
 #define TC6_LIB_VER_BUGFIX (6U)
-#define TC6_LIB_VER_STRING "V3.1.6"
+#define TC6_LIB_VER_STRING "V3.1.7"
 
 struct TC6_t;
 typedef struct TC6_t TC6_t;